refactor: deepen post-action observation
